trial

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -9,8 +9,8 @@
 int _atoi(char *s)
 {
 	int b, a, min, digit;
-	int num = 0;
-	int result;
+	int64_t num = 0;
+	int64_t result;
 
 	b = 0;
 	while (*(s + b) != '\0')
